Move CanUpgradeLedgerTables to Ouroboros.Consensus.Ledger.Tables
